+++ g2p_programs/models/__init__.py
@@ -28,3 +28,4 @@ from . import cycle
 from . import cycle_membership
 from . import voucher
 from . import cash_voucher
+from . import accounting