+++ g2p_programs/models/accounting/__init__.py
@@ -17,7 +17,4 @@
 # limitations under the License.
 #
 
-from . import models
-from . import wizard
-
-# from . import controllers
+from . import account_journal